+++ packages/spry/lib/constants.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: constant_identifier_names
+
+/// Spry core constants.
+library spry.constants;
+
+/// Origin http request in [Context] store under this key.
+///
+/// See [Context.get] and [Context.set].
+const Symbol SPRY_HTTP_REQUEST = Symbol('#spry.http-request');