+++ packages/spry/lib/extension.dart
@@ -1,0 +1,4 @@
+/// Spry core extension.
+library spry.extension;
+
+export 'src/extensions/middleware_extension.dart';