+++ packages/spry/lib/src/constants.dart
@@ -1,4 +1,0 @@
-/// Origin http request in [Context] store under this key.
-///
-/// See [Context.get] and [Context.set].
-const Symbol spryHttpRequest = Symbol('#spry.http-request');