+++ packages/spry_router/example/spry_router_example.dart
@@ -1,0 +1,6 @@
+import 'package:spry_router/spry_router.dart';
+
+void main() {
+  var awesome = Awesome();
+  print('awesome: ${awesome.isAwesome}');
+}