+++ packages/spry_router/lib/spry_router.dart
@@ -1,0 +1,1 @@
+library spry.router;