+++ packages/spry/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+- Move `spryHttpRequest` to `spry.constants` library.
+- Add moddleware extension.
+
 ## 0.0.3
 
 - Public `HttpRequest` store key.